backend/requirements.txt
python-dotenv==1.0.1

flask_pymongo==2.3.0
flask==3.0.2
flask_jwt_extended==4.6.0
flask_cors==4.0.1
flask_restful==0.3.10
requests==2.32.3
requests-toolbelt==1.0.0

Werkzeug==3.0.3
pillow==10.3.0
cryptography==42.0.8

black==24.10.0
isort==5.13.2
ruff==0.8.0
mypy==1.13.0

types-pytz==2024.2.0.20241003
types-flask-cors==5.0.0.20240902
types-requests==2.32.0.20241016